requirements.txt
pandas
boto3
nltk
dask
pyspark
openai
transformers
sentence-transformers
faiss-cpu
matplotlib
fastapi
uvicorn